+++ src/index.ts
@@ -1,9 +1,11 @@
 import { PDFExtract, PDFExtractOptions, PDFExtractResult } from 'pdf.js-extract'
 
-import { ISyllabusCourse } from './types'
+import { ICourse, ISyllabusCourse } from './types'
 import { cleanGeneralInfo } from './utils'
 
 // Key words from the syllabus main page
+const UNI = process.env.UNI as string
+const NONE = process.env.NONE as string
 const SYLLABUS = process.env.SYLLABUS as string
 const GENERAL_INFO = process.env.GENERAL_INFO as string
 const SOMMELIER = process.env.SOMMELIER as string
@@ -23,10 +25,7 @@ const THEORY = process.env.THEORY as string
 const LABORATORY = process.env.LABORATORY as string
 
 const pdfExtract = new PDFExtract()
-const options: PDFExtractOptions = {
-  firstPage: 18,
-  lastPage : 21
-}
+const options: PDFExtractOptions = {}
 
 // const fullContent: ISyllabusCourse[] = []
 
@@ -38,177 +37,75 @@ const extractData = async () => {
     )
     const { pages } = data
 
+    let planContent = ''
     pages.forEach(({ content }) => {
-      const currentText: string[] = []
-
       content.forEach(text => {
         if (text.str !== '' && text.str !== ' ')
-          currentText.push(text.str)
+        planContent += text.str
       })
-
-      const isTheFirstPage = currentText.indexOf(SYLLABUS) !== -1
-      const pageContent: ISyllabusCourse = {
-        analyticProgram: [],
-        competencies   : [],
-        generalInfo    : {
-          condition: '',
-          course   : {
-            code: '',
-            name: ''
-          },
-          credits         : 0,
-          evaluationSystem: '',
-          hoursPerWeek    : {
-            laboratory: 0,
-            theory    : 0,
-            total     : 0
-          },
-          preRequirements: {
-            code: '',
-            name: ''
-          }
-        },
-        sommelier: ''
-      }
-
-      if (isTheFirstPage) {
-        // Removing everything before the syllabus
-        const positionOfSyllabus = currentText.indexOf(SYLLABUS)
-        currentText.splice(0, positionOfSyllabus + 1)
-
-        // Removing "Información General"
-        const positionOfGeneralInfo = currentText.indexOf(GENERAL_INFO)
-        currentText.splice(0, positionOfGeneralInfo + 1)
-
-        // Saving the sommelier
-        let positionOfSommelier = currentText.indexOf(SOMMELIER)
-
-        // Saving general info and replacing REQUIREMENTS by REQUIREMENT in case
-        // the present word is REQUIREMENTS
-        const generalInfo = currentText
-          .splice(0, positionOfSommelier)
-          .join(' ')
-          .replace(`${REQUIREMENT}S`, REQUIREMENT)
-        const preRequirements = cleanGeneralInfo({
-          positionEndSlice  : generalInfo.indexOf(CONDITION),
-          positionStartSlice: generalInfo
-            .indexOf(PRE_REQUIREMENTS) + PRE_REQUIREMENTS.length,
-          text: generalInfo
-        })
-        const hoursPerWeek = cleanGeneralInfo({
-          positionEndSlice  : generalInfo.indexOf(EVALUATION_SYSTEM),
-          positionStartSlice: generalInfo
-            .indexOf(HOURS_PER_WEEK) + HOURS_PER_WEEK.length,
-          text: generalInfo
-        })
-        pageContent.generalInfo = {
-          ...pageContent.generalInfo,
-          condition: cleanGeneralInfo({
-            positionEndSlice  : generalInfo.indexOf(HOURS_PER_WEEK),
-            positionStartSlice: generalInfo
-              .indexOf(CONDITION) + CONDITION.length,
-            text: generalInfo
-          }),
-          course: {
-            code: cleanGeneralInfo({
-              positionEndSlice  : generalInfo.indexOf(CREDITS),
-              positionStartSlice: generalInfo.indexOf(CODE) + CODE.length,
-              text              : generalInfo
-            }).replace(' ', ''),
-            name: cleanGeneralInfo({
-              positionEndSlice  : generalInfo.indexOf(CODE),
-              positionStartSlice: generalInfo.indexOf(COURSE) + COURSE.length,
-              text              : generalInfo
-            })
-          },
-          credits: parseInt(
-            cleanGeneralInfo({
-              positionEndSlice  : generalInfo.indexOf(PRE_REQUIREMENTS),
-              positionStartSlice: generalInfo.indexOf(CREDITS) + CREDITS.length,
-              text              : generalInfo
-            }
-          )),
-          evaluationSystem: cleanGeneralInfo({
-            positionStartSlice: generalInfo
-              .indexOf(EVALUATION_SYSTEM) + EVALUATION_SYSTEM.length,
-            text: generalInfo
-          }),
-          hoursPerWeek: {
-            laboratory: parseInt(
-              hoursPerWeek
-                .slice(
-                  hoursPerWeek.indexOf(LABORATORY) + LABORATORY.length,
-                  hoursPerWeek.indexOf(')')
-                )
-                .replace('0', '')
-                .replace(/ /g, '')
-            ),
-            theory: parseInt(
-              hoursPerWeek
-                .slice(
-                  hoursPerWeek.indexOf(THEORY) + THEORY.length,
-                  hoursPerWeek.indexOf(',')
-                )
-                .replace('0', '')
-                .replace(/ /g, '')
-            ),
-            total: parseInt(
-              hoursPerWeek
-                .slice(0, hoursPerWeek.indexOf('('))
-                .replace('0', '')
-                .replace(/ /g, '')
-            )
-          },
-          preRequirements: {
-            code: preRequirements.slice(0, preRequirements.indexOf(' –')),
-            name: preRequirements.slice(preRequirements.indexOf('– ') + 2)
-          }
-        }
-
-        // Updating sommelier position after splice
-        positionOfSommelier = currentText.indexOf(SOMMELIER)
-
-        // Removing "Sumilla"
-        currentText.splice(0, positionOfSommelier + 1)
-        let positionOfCompetencies = currentText.indexOf(COMPETENCIES)
-
-        // Verifying if the competencies are in this page
-        if (positionOfCompetencies !== -1) {
-          // Saving sommelier
-          pageContent.sommelier = currentText
-            .splice(0, positionOfCompetencies)
-            .join(' ')
-            .replace(/  +/g, ' ')
-            .trim()
-          // Updating competencies position
-          positionOfCompetencies = currentText.indexOf(COMPETENCIES)
-          // Removing "Competencias" and the first bullet
-          currentText.splice(0, positionOfCompetencies + 2)
-
-          // If the competencies are still here, I save them
-          if (currentText.length > 0) {
-            let competence = ''
-            currentText.forEach((e, index, arr) => {
-              if (e !== BULLET) {
-                competence += e
-                if (index === arr.length - 1)
-                  pageContent.competencies.push(
-                    competence.replace(/  +/g, ' ').trim()
-                  )
-              } else {
-                pageContent.competencies.push(
-                  competence.replace(/  +/g, ' ').trim()
-                )
-                competence = ''
-              }
-            })
-          }
-        } else
-          pageContent.sommelier = currentText.splice(0).join(' ')
-
-        console.log(JSON.stringify(pageContent, null, 2))
-      }
     })
+
+    const courseContentArray = planContent.split(UNI)
+    courseContentArray.shift()
+
+    courseContentArray.forEach(courseContent => {
+      const courseName = cleanGeneralInfo({
+        positionEndSlice  : courseContent.indexOf(CODE),
+        positionStartSlice: courseContent.indexOf(COURSE) + COURSE.length,
+        text              : courseContent
+      })
+      const courseCode = cleanGeneralInfo({
+        positionEndSlice  : courseContent.indexOf(CREDITS),
+        positionStartSlice: courseContent.indexOf(CODE) + CODE.length,
+        text              : courseContent
+      })
+      const credits = parseInt(
+        cleanGeneralInfo({
+          positionEndSlice  : courseContent.indexOf('('),
+          positionStartSlice: courseContent.indexOf(CREDITS) + CREDITS.length,
+          text              : courseContent
+        }).trim()
+      )
+      let preRequirementsAux = cleanGeneralInfo({
+        positionEndSlice  : courseContent.indexOf(CONDITION),
+        positionStartSlice: courseContent
+          .indexOf(PRE_REQUIREMENTS) + PRE_REQUIREMENTS.length,
+        text: courseContent
+      })
+      let preRequirements: string[] | null = null
+      let finalPreRequirements: ICourse[] | null = null
+
+      if (!preRequirementsAux.includes(NONE)) {
+        if (preRequirementsAux[0] === '-')
+          preRequirementsAux = preRequirementsAux.slice(1)
+
+        preRequirements = preRequirementsAux
+          .replace(/ - /g, '/')
+          .split('-')
+
+        const auxFinalPreRequirements: ICourse[] = []
+
+        preRequirements.forEach(preRequirement => {
+          const aux = preRequirement.split('/')
+
+          auxFinalPreRequirements.push({
+            code: aux[0],
+            name: aux[1]
+          })
+        })
+
+        finalPreRequirements = auxFinalPreRequirements
+      }
+
+      console.log({
+        courseCode,
+        courseName,
+        credits,
+        preRequirements: finalPreRequirements
+      })
+    })
+
+    // console.log({ courseContentArray: courseContentArray.splice(0, 3) })
   } catch (error) {
     console.error(error)
   }